Fixed FloatValidator conditioner
git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@1028 0d8f46c8-9b30-49d0-b470-0f6283dc92e8

+++ lib/certify.py
@@ -351,8 +351,8 @@ class FloatValidator(NumericalValidator):
             raise
 
     def condition(self, value):
-        val = int(value)
-        if val == -999:
+        val = float(value)
+        if val == -999.0:
             val = '*'
         return val
